05. Proyecto_Angular/backend/requirements.txt
django
django-environ
mysqlclient
djangorestframework
django-cors-headers